add response show response changed, and delete and add meal all works good

+++ client/src/main/java/il/cshaifasweng/OCSFMediatorExample/client/SimpleClient.java
@@ -20,7 +20,7 @@ public class SimpleClient extends AbstractClient {
 	public static int Port = 3000;
 	private static UserCheck UserClient = null;
 	private static boolean logged = false;
-	public static int deliveryPrice = 25;
+	public static double deliveryPrice = 25;
 	public static RestaurantList restaurantList = null;
 
 	private SimpleClient(String host, int port) {
